Release v.1.1
Добавлен режим отладки, добавлена проверка корректности значений в
EEPROM при загрузке.
Исправлено время прогрева.
Изменены калибровочные данные.

+++ CO2meter/CO2meter/Display.h
@@ -11,7 +11,7 @@
 
 void DisplayTimer0Init();
 
-extern uint8_t displayValue;
+extern uint16_t displayValue;
 extern uint8_t displayMode;
 
 #endif /* DISPLAY_H_ */
